Separating problem definition and enums into separate files
